add useHeadset hook
