+++ .changeset/lucky-ears-applaud.md
@@ -1,0 +1,5 @@
+---
+'@threlte/xr': patch
+---
+
+Add useHeadset hook
