+++ apps/docs/src/content/reference/xr/headset.mdx
@@ -22,7 +22,9 @@
 </Headset>
 ```
 
-Like a Portal, it can exist anywhere in your Threlte application.
+If you need to only read from the current headset pose, the [useHeadset](/docs/reference/xr/use-headset) hook is available.
+
+Like a [Portal](/docs/reference/extras/portal), you can place it anywhere in your Threlte application.
 
 `<Headset>` will sync position and rotation with the current camera when not in an immersive XR session.
